More traits usage
