+++ lib/ApiOperations/All.php
@@ -5,7 +5,7 @@ namespace Stripe\ApiOperations;
 /**
  * Trait for listable resources. Adds a `all()` static method to the class.
  *
- * This trait should only be applied to classes that derive from ApiResource.
+ * This trait should only be applied to classes that derive from StripeObject.
  */
 trait All
 {
@@ -17,6 +17,18 @@ trait All
      */
     public static function all($params = null, $opts = null)
     {
-        return self::_all($params, $opts);
+        self::_validateParams($params);
+        $url = static::classUrl();
+
+        list($response, $opts) = static::_staticRequest('get', $url, $params, $opts);
+        $obj = \Stripe\Util\Util::convertToStripeObject($response->json, $opts);
+        if (!is_a($obj, 'Stripe\\Collection')) {
+            $class = get_class($obj);
+            $message = "Expected type \"Stripe\\Collection\", got \"$class\" instead";
+            throw new Error\Api($message);
+        }
+        $obj->setLastResponse($response);
+        $obj->setRequestParams($params);
+        return $obj;
     }
 }
