+++ lib/ApiOperations/Create.php
@@ -5,7 +5,7 @@ namespace Stripe\ApiOperations;
 /**
  * Trait for creatable resources. Adds a `create()` static method to the class.
  *
- * This trait should only be applied to classes that derive from ApiResource.
+ * This trait should only be applied to classes that derive from StripeObject.
  */
 trait Create
 {
@@ -17,6 +17,12 @@ trait Create
      */
     public static function create($params = null, $options = null)
     {
-        return self::_create($params, $options);
+        self::_validateParams($params);
+        $url = static::classUrl();
+
+        list($response, $opts) = static::_staticRequest('post', $url, $params, $options);
+        $obj = \Stripe\Util\Util::convertToStripeObject($response->json, $opts);
+        $obj->setLastResponse($response);
+        return $obj;
     }
 }
