+++ lib/ApiOperations/Delete.php
@@ -5,7 +5,7 @@ namespace Stripe\ApiOperations;
 /**
  * Trait for deletable resources. Adds a `delete()` method to the class.
  *
- * This trait should only be applied to classes that derive from ApiResource.
+ * This trait should only be applied to classes that derive from StripeObject.
  */
 trait Delete
 {
@@ -17,6 +17,11 @@ trait Delete
      */
     public function delete($params = null, $opts = null)
     {
-        return $this->_delete($params, $opts);
+        self::_validateParams($params);
+
+        $url = $this->instanceUrl();
+        list($response, $opts) = $this->_request('delete', $url, $params, $opts);
+        $this->refreshFrom($response, $opts);
+        return $this;
     }
 }
