+++ lib/ApiOperations/Request.php
@@ -1,0 +1,64 @@
+<?php
+
+namespace Stripe\ApiOperations;
+
+/**
+ * Trait for resources that need to make API requests.
+ *
+ * This trait should only be applied to classes that derive from StripeObject.
+ */
+trait Request
+{
+    /**
+     * @param array|null|mixed $params The list of parameters to validate
+     *
+     * @throws Stripe\Error\Api if $params exists and is not an array
+     */
+    protected static function _validateParams($params = null)
+    {
+        if ($params && !is_array($params)) {
+            $message = "You must pass an array as the first argument to Stripe API "
+               . "method calls.  (HINT: an example call to create a charge "
+               . "would be: \"Stripe\\Charge::create(['amount' => 100, "
+               . "'currency' => 'usd', 'source' => 'tok_1234'])\")";
+            throw new Error\Api($message);
+        }
+    }
+
+    /**
+     * @param string $method HTTP method ('get', 'post', etc.)
+     * @param string $url URL for the request
+     * @param array $params list of parameters for the request
+     * @param array|string|null $options
+     *
+     * @return array tuple containing (the JSON response, $options)
+     */
+    protected function _request($method, $url, $params = [], $options = null)
+    {
+        $opts = $this->_opts->merge($options);
+        list($resp, $options) = static::_staticRequest($method, $url, $params, $opts);
+        $this->setLastResponse($resp);
+        return [$resp->json, $options];
+    }
+
+    /**
+     * @param string $method HTTP method ('get', 'post', etc.)
+     * @param string $url URL for the request
+     * @param array $params list of parameters for the request
+     * @param array|string|null $options
+     *
+     * @return array tuple containing (the JSON response, $options)
+     */
+    protected static function _staticRequest($method, $url, $params, $options)
+    {
+        $opts = \Stripe\Util\RequestOptions::parse($options);
+        $requestor = new \Stripe\ApiRequestor($opts->apiKey, static::baseUrl());
+        list($response, $opts->apiKey) = $requestor->request($method, $url, $params, $opts->headers);
+        foreach ($opts->headers as $k => $v) {
+            if (!array_key_exists($k, self::$HEADERS_TO_PERSIST)) {
+                unset($opts->headers[$k]);
+            }
+        }
+        return [$response, $opts];
+    }
+}
