+++ lib/ApiOperations/Update.php
@@ -6,29 +6,41 @@ namespace Stripe\ApiOperations;
  * Trait for updatable resources. Adds an `update()` static method and a
  * `save()` method to the class.
  *
- * This trait should only be applied to classes that derive from ApiResource.
+ * This trait should only be applied to classes that derive from StripeObject.
  */
 trait Update
 {
     /**
      * @param string $id The ID of the resource to update.
      * @param array|null $params
-     * @param array|string|null $options
+     * @param array|string|null $opts
      *
      * @return ApiResource The updated resource.
      */
-    public static function update($id, $params = null, $options = null)
+    public static function update($id, $params = null, $opts = null)
     {
-        return self::_update($id, $params, $options);
+        self::_validateParams($params);
+        $url = static::resourceUrl($id);
+
+        list($response, $opts) = static::_staticRequest('post', $url, $params, $opts);
+        $obj = \Stripe\Util\Util::convertToStripeObject($response->json, $opts);
+        $obj->setLastResponse($response);
+        return $obj;
     }
 
     /**
-     * @param array|string|null $options
+     * @param array|string|null $opts
      *
      * @return ApiResource The saved resource.
      */
-    public function save($options = null)
+    public function save($opts = null)
     {
-        return $this->_save($options);
+        $params = $this->serializeParameters();
+        if (count($params) > 0) {
+            $url = $this->instanceUrl();
+            list($response, $opts) = $this->_request('post', $url, $params, $opts);
+            $this->refreshFrom($response, $opts);
+        }
+        return $this;
     }
 }
