+++ lib/ApiResource.php
@@ -9,6 +9,8 @@ namespace Stripe;
  */
 abstract class ApiResource extends StripeObject
 {
+    use ApiOperations\Request;
+
     /**
      * @return Stripe\Util\Set A list of fields that can be their own type of
      * API resource (say a nested card under an account for example), and if
@@ -136,267 +138,5 @@ abstract class ApiResource extends StripeObject
     public function instanceUrl()
     {
         return static::resourceUrl($this['id']);
-    }
-
-    /**
-     * @param array|null|mixed $params The list of parameters to validate
-     *
-     * @throws Stripe\Error\Api if $params exists and is not an array
-     */
-    protected static function _validateParams($params = null)
-    {
-        if ($params && !is_array($params)) {
-            $message = "You must pass an array as the first argument to Stripe API "
-               . "method calls.  (HINT: an example call to create a charge "
-               . "would be: \"Stripe\\Charge::create(['amount' => 100, "
-               . "'currency' => 'usd', 'source' => 'tok_1234'])\")";
-            throw new Error\Api($message);
-        }
-    }
-
-    /**
-     * @param string $method HTTP method ('get', 'post', etc.)
-     * @param string $url URL for the request
-     * @param array $params list of parameters for the request
-     * @param array|string|null $options
-     *
-     * @return array tuple containing (the JSON response, $options)
-     */
-    protected function _request($method, $url, $params = [], $options = null)
-    {
-        $opts = $this->_opts->merge($options);
-        list($resp, $options) = static::_staticRequest($method, $url, $params, $opts);
-        $this->setLastResponse($resp);
-        return [$resp->json, $options];
-    }
-
-    /**
-     * @param string $method HTTP method ('get', 'post', etc.)
-     * @param string $url URL for the request
-     * @param array $params list of parameters for the request
-     * @param array|string|null $options
-     *
-     * @return array tuple containing (the JSON response, $options)
-     */
-    protected static function _staticRequest($method, $url, $params, $options)
-    {
-        $opts = Util\RequestOptions::parse($options);
-        $requestor = new ApiRequestor($opts->apiKey, static::baseUrl());
-        list($response, $opts->apiKey) = $requestor->request($method, $url, $params, $opts->headers);
-        foreach ($opts->headers as $k => $v) {
-            if (!array_key_exists($k, self::$HEADERS_TO_PERSIST)) {
-                unset($opts->headers[$k]);
-            }
-        }
-        return [$response, $opts];
-    }
-
-    /**
-     * @param string|array $id The ID of the API resource to retrieve, or the
-     *   list of parameters for the request.
-     * @param array|string|null $options
-     *
-     * @return Stripe\ApiResource the retrieved API resource
-     */
-    protected static function _retrieve($id, $options = null)
-    {
-        $opts = Util\RequestOptions::parse($options);
-        $instance = new static($id, $opts);
-        $instance->refresh();
-        return $instance;
-    }
-
-    /**
-     * @param array $params The list of parameters for the request.
-     * @param array|string|null $options
-     *
-     * @return Stripe\Collection the retrieved list of API resources
-     */
-    protected static function _all($params = null, $options = null)
-    {
-        self::_validateParams($params);
-        $url = static::classUrl();
-
-        list($response, $opts) = static::_staticRequest('get', $url, $params, $options);
-        $obj = Util\Util::convertToStripeObject($response->json, $opts);
-        if (!is_a($obj, 'Stripe\\Collection')) {
-            $class = get_class($obj);
-            $message = "Expected type \"Stripe\\Collection\", got \"$class\" instead";
-            throw new Error\Api($message);
-        }
-        $obj->setLastResponse($response);
-        $obj->setRequestParams($params);
-        return $obj;
-    }
-
-    /**
-     * @param array $params The list of parameters for the request.
-     * @param array|string|null $options
-     *
-     * @return Stripe\ApiResource the created API resource
-     */
-    protected static function _create($params = null, $options = null)
-    {
-        self::_validateParams($params);
-        $url = static::classUrl();
-
-        list($response, $opts) = static::_staticRequest('post', $url, $params, $options);
-        $obj = Util\Util::convertToStripeObject($response->json, $opts);
-        $obj->setLastResponse($response);
-        return $obj;
-    }
-
-    /**
-     * @param string $id The ID of the API resource to update.
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return ApiResource the updated API resource
-     */
-    protected static function _update($id, $params = null, $options = null)
-    {
-        self::_validateParams($params);
-        $url = static::resourceUrl($id);
-
-        list($response, $opts) = static::_staticRequest('post', $url, $params, $options);
-        $obj = Util\Util::convertToStripeObject($response->json, $opts);
-        $obj->setLastResponse($response);
-        return $obj;
-    }
-
-    /**
-     * @param array|string|null $options
-     *
-     * @return Stripe\ApiResource the updated API resource
-     */
-    protected function _save($options = null)
-    {
-        $params = $this->serializeParameters();
-        if (count($params) > 0) {
-            $url = $this->instanceUrl();
-            list($response, $opts) = $this->_request('post', $url, $params, $options);
-            $this->refreshFrom($response, $opts);
-        }
-        return $this;
-    }
-
-    /**
-     * @param array $params The list of parameters for the request.
-     * @param array|string|null $options
-     *
-     * @return Stripe\ApiResource the deleted API resource
-     */
-    protected function _delete($params = null, $options = null)
-    {
-        self::_validateParams($params);
-
-        $url = $this->instanceUrl();
-        list($response, $opts) = $this->_request('delete', $url, $params, $options);
-        $this->refreshFrom($response, $opts);
-        return $this;
-    }
-
-    /**
-     * @param string $method
-     * @param string $url
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Stripe\StripeObject
-     */
-    protected static function _nestedResourceOperation($method, $url, $params = null, $options = null)
-    {
-        self::_validateParams($params);
-
-        list($response, $opts) = static::_staticRequest($method, $url, $params, $options);
-        $obj = Util\Util::convertToStripeObject($response->json, $opts);
-        $obj->setLastResponse($response);
-        return $obj;
-    }
-
-    /**
-     * @param string $id
-     * @param string $nestedPath
-     * @param string|null $nestedId
-     *
-     * @return string
-     */
-    protected static function _nestedResourceUrl($id, $nestedPath, $nestedId = null)
-    {
-        $url = static::resourceUrl($id) . $nestedPath;
-        if ($nestedId !== null) {
-            $url .= "/$nestedId";
-        }
-        return $url;
-    }
-
-    /**
-     * @param string $id
-     * @param string $nestedPath
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Stripe\StripeObject
-     */
-    protected static function _createNestedResource($id, $nestedPath, $params = null, $options = null)
-    {
-        $url = static::_nestedResourceUrl($id, $nestedPath);
-        return self::_nestedResourceOperation('post', $url, $params, $options);
-    }
-
-    /**
-     * @param string $id
-     * @param string $nestedPath
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Stripe\StripeObject
-     */
-    protected static function _retrieveNestedResource($id, $nestedPath, $nestedId, $params = null, $options = null)
-    {
-        $url = static::_nestedResourceUrl($id, $nestedPath, $nestedId);
-        return self::_nestedResourceOperation('get', $url, $params, $options);
-    }
-
-    /**
-     * @param string $id
-     * @param string $nestedPath
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Stripe\StripeObject
-     */
-    protected static function _updateNestedResource($id, $nestedPath, $nestedId, $params = null, $options = null)
-    {
-        $url = static::_nestedResourceUrl($id, $nestedPath, $nestedId);
-        return self::_nestedResourceOperation('post', $url, $params, $options);
-    }
-
-    /**
-     * @param string $id
-     * @param string $nestedPath
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Stripe\StripeObject
-     */
-    protected static function _deleteNestedResource($id, $nestedPath, $nestedId, $params = null, $options = null)
-    {
-        $url = static::_nestedResourceUrl($id, $nestedPath, $nestedId);
-        return self::_nestedResourceOperation('delete', $url, $params, $options);
-    }
-
-    /**
-     * @param string $id
-     * @param string $nestedPath
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Stripe\StripeObject
-     */
-    protected static function _allNestedResources($id, $nestedPath, $params = null, $options = null)
-    {
-        $url = static::_nestedResourceUrl($id, $nestedPath);
-        return self::_nestedResourceOperation('get', $url, $params, $options);
     }
 }
