+++ lib/ApplicationFee.php
@@ -10,6 +10,7 @@ namespace Stripe;
 class ApplicationFee extends ApiResource
 {
     use ApiOperations\All;
+    use ApiOperations\NestedResource;
     use ApiOperations\Retrieve;
 
     const PATH_REFUNDS = '/refunds';
