+++ lib/ApplicationFeeRefund.php
@@ -9,6 +9,10 @@ namespace Stripe;
  */
 class ApplicationFeeRefund extends ApiResource
 {
+    use ApiOperations\Update {
+        save as protected _save;
+    }
+
     /**
      * @return string The API URL for this Stripe refund.
      */
