+++ lib/Subscription.php
@@ -11,6 +11,9 @@ class Subscription extends ApiResource
 {
     use ApiOperations\All;
     use ApiOperations\Create;
+    use ApiOperations\Delete {
+        delete as protected _delete;
+    }
     use ApiOperations\Retrieve;
     use ApiOperations\Update;
