+++ lib/Transfer.php
@@ -27,6 +27,7 @@ class Transfer extends ApiResource
 {
     use ApiOperations\All;
     use ApiOperations\Create;
+    use ApiOperations\NestedResource;
     use ApiOperations\Retrieve;
     use ApiOperations\Update;
